Assignment of Lesson 5.1
